+++ package.json
@@ -15,7 +15,8 @@
     "@tailwindcss/vite": "^4.0.6",
     "framer-motion": "^12.4.3",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.5.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
